requirements.txt
altgraph==0.17
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.10
auto-py-to-exe==2.7.8
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.4.5.1
cffi==1.14.2
chardet==3.0.4
cryptography==3.3.1
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.1
django-allauth==0.42.0
django-bootstrap-form==3.4
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-embed-video==1.3.3
django-filter==2.3.0
django-js-asset==1.2.2
django-service-objects==0.7.0
django-tinymce==3.0.2
docutils==0.16
Eel==0.12.4
filelock==3.0.12
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
Kivy==1.11.1
kivy-deps.glew==0.2.0
kivy-deps.sdl2==0.2.0
Kivy-Garden==0.1.4
MouseInfo==0.1.3
numpy==1.18.4
oauthlib==3.1.0
pefile==2019.4.18
Pillow==7.1.2
psycopg2-binary==2.8.6
PyAutoGUI==0.9.50
pycparser==2.20
pygame==1.9.6
PyGetWindow==0.0.8
Pygments==2.6.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
PyMsgBox==1.0.8
pyperclip==1.8.0
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
pyTelegramBotAPI==3.7.1
python-telegram-bot==13.1
python3-openid==3.2.0
PyTweening==1.0.3
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
tornado==6.1
tzlocal==2.1
urllib3==1.25.9
virtualenv==20.0.21
waitress==1.4.4
whichcraft==0.6.1
whitenoise==5.2.0
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.1.0
